Allow non-modifiable credentials stores
