+++ src/main/java/com/cloudbees/plugins/credentials/ModifiableDomainsCredentialsStore.java
@@ -1,0 +1,40 @@
+/*
+ * The MIT License
+ *
+ * Copyright (c) 2013-2016, CloudBees, Inc., Stephen Connolly.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.cloudbees.plugins.credentials;
+
+import hudson.model.Saveable;
+import hudson.security.AccessControlled;
+
+/**
+ * A store of {@link Credentials}. Each {@link ModifiableDomainsCredentialsStore} is associated with one and only one
+ * {@link CredentialsProvider} though a {@link CredentialsProvider} may provide multiple {@link ModifiableDomainsCredentialsStore}s
+ * (for example a folder scoped {@link CredentialsProvider} may provide a {@link ModifiableDomainsCredentialsStore} for each folder
+ * or a user scoped {@link CredentialsProvider} may provide a {@link ModifiableDomainsCredentialsStore} for each user).
+ *
+ * @author Stephen Connolly
+ * @since 1.8
+ */
+interface ModifiableDomainsCredentialsStore extends AccessControlled, Saveable {
+
+}
