+++ src/main/resources/com/cloudbees/plugins/credentials/CredentialsSelectHelper/WrappedCredentialsStore/dialog.jelly
@@ -30,32 +30,34 @@
         <l:icon class="${it.iconClassName} icon-xlg"/> ${it.description}: ${it.displayName}
       </h2>
     </div>
-    <div class="bd">
-      <h3>
-      <l:icon class="icon-credentials-new-credential icon-md"/>
-      ${%Add Credentials}
-      </h3>
-      <form action="${it.url}/addCredentials" method="POST" id="credentials-dialog-form">
-        <table width="100%">
-          <f:entry title="${%Domain}">
-            <select class="setting-input" name="_.domain">
-              <j:forEach var="domain" items="${it.wrappers.entrySet()}">
-                <f:option value="${domain.key}" selected="${domain.key=='_'}">${domain.value.displayName}</f:option>
-              </j:forEach>
-            </select>
-          </f:entry>
-          <f:block>
-            <j:set var="descriptors" value="${it.credentialsDescriptors}"/>
-            <st:include page="credential"/>
-          </f:block>
-          <f:block>
-            <input id="credentials-add-submit" class="yui-button" value="${%Add}"
-                   onclick="return window.credentials.addSubmit(this);"/>
-            <input id="credentials-add-abort" class="yui-button"
-                   onclick="window.credentials.dialog.hide(); return false;" value="${%Cancel}"/>
-          </f:block>
-        </table>
-      </form>
-    </div>
+    <j:if test="${it.credentialsModifiable}">
+      <div class="bd">
+        <h3>
+        <l:icon class="icon-credentials-new-credential icon-md"/>
+        ${%Add Credentials}
+        </h3>
+        <form action="${it.url}/addCredentials" method="POST" id="credentials-dialog-form">
+          <table width="100%">
+            <f:entry title="${%Domain}">
+              <select class="setting-input" name="_.domain">
+                <j:forEach var="domain" items="${it.wrappers.entrySet()}">
+                  <f:option value="${domain.key}" selected="${domain.key=='_'}">${domain.value.displayName}</f:option>
+                </j:forEach>
+              </select>
+            </f:entry>
+            <f:block>
+              <j:set var="descriptors" value="${it.credentialsDescriptors}"/>
+              <st:include page="credential"/>
+            </f:block>
+            <f:block>
+              <input id="credentials-add-submit" class="yui-button" value="${%Add}"
+                     onclick="return window.credentials.addSubmit(this);"/>
+              <input id="credentials-add-abort" class="yui-button"
+                     onclick="window.credentials.dialog.hide(); return false;" value="${%Cancel}"/>
+            </f:block>
+          </table>
+        </form>
+      </div>
+    </j:if>
   </l:ajax>
 </j:jelly>
