+++ src/main/resources/com/cloudbees/plugins/credentials/CredentialsStoreAction/CredentialsWrapper/move.jelly
@@ -1,31 +1,38 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:i="jelly:fmt">
   <l:layout title="${it.fullDisplayName}" norefresh="true" permission="${it.domain.parent.DELETE}">
-		<st:include page="sidepanel.jelly" />
-		<l:main-panel>
+    <st:include page="sidepanel.jelly" />
+    <l:main-panel>
       <h1><img src="${resURL}/plugin/credentials/images/48x48/move.png" alt=""/>${%Move}</h1>
-      <form method="post" action="doMove">
-        ${%blurb(it.displayName,it.typeName)}
-        <br />
-        <select name="destination" class="select setting-input" style="width:auto;">
-          <j:forEach var="destination" items="${it.domain.parent.domains.values()}">
-            <j:choose>
-              <j:when test="${destination.domain == it.domain.domain}">
-                <option selected="yes" value="${destination.fullName}">
+      <j:choose>
+        <j:when test="${it.credentialsModifiable}">
+          <form method="post" action="doMove">
+            ${%blurb(it.displayName,it.typeName)}
+            <br />
+            <select name="destination" class="select setting-input" style="width:auto;">
+              <j:forEach var="destination" items="${it.domain.parent.domains.values()}">
+                <j:choose>
+                  <j:when test="${destination.domain == it.domain.domain}">
+                    <option selected="yes" value="${destination.fullName}">
                   ${app.displayName} &#187; ${destination.fullDisplayName}
-                </option>
-              </j:when>
-              <j:otherwise>
-                <option value="${destination.fullName}">
+                    </option>
+                  </j:when>
+                  <j:otherwise>
+                    <option value="${destination.fullName}">
                   ${app.displayName} &#187; ${destination.fullDisplayName}
-                </option>
-              </j:otherwise>
-            </j:choose>
-          </j:forEach>
-        </select>
-        <br />
-        <f:submit value="${%Move}"/>
-      </form>
+                    </option>
+                  </j:otherwise>
+                </j:choose>
+              </j:forEach>
+            </select>
+            <br />
+            <f:submit value="${%Move}"/>
+        </form>
+      </j:when>
+      <j:otherwise>
+        <p>The credentials provider is not modifiable and therefore cannot be moved</p>
+      </j:otherwise>
+    </j:choose>
     </l:main-panel>
-	</l:layout>
+  </l:layout>
 </j:jelly>
