+++ src/main/resources/com/cloudbees/plugins/credentials/CredentialsStoreAction/DomainWrapper/index.jelly
@@ -31,9 +31,19 @@
       <h1><l:icon class="icon-credentials-domain icon-xlg"/>${it.displayName}</h1>
       <div>
         <j:out value="${it.description!=null ? app.markupFormatter.translate(it.description) : ''}" />
+        <j:if test="${!it.store.credentialsModifiable}">
+          <br />This is a read-only credentials provider. Please see the docs to find out how to add/edit/remove credentials
+        </j:if>
       </div>
       <t:setIconSize/>
       <table class="sortable pane bigtable">
+        <colgroup>
+          <col width="5%" />
+          <col width="20%" />
+          <col width="20%" />
+          <col width="50%;" />
+          <col width="5%;" />
+        </colgroup>
         <tr style="border-top: 0px;">
           <th>
             <st:nbsp/>
@@ -58,7 +68,7 @@
             <tr>
                <td colspan="5" align="center">
                  <j:choose>
-                   <j:when test="${it.store.hasPermission(it.parent.CREATE)}">
+                   <j:when test="${it.store.credentialsModifiable and it.store.hasPermission(it.parent.CREATE)}">
                      ${%noCredentialsAddSome}
                    </j:when>
                    <j:otherwise>
@@ -87,7 +97,7 @@
                   ${safeDescription}
                 </td>
                 <td>
-                  <j:if test="${h.hasPermission(c, it.parent.UPDATE)}">
+                  <j:if test="${it.credentialsModifiable and h.hasPermission(c, it.parent.UPDATE)}">
                     <a href="credential/${c.urlName}/update">
                       <img src="${resURL}/images/${iconSize}/setting.png" alt="${%Update}"
                            tooltip="${%Update}" class="icon${iconSize}"/>
