+++ src/main/resources/com/cloudbees/plugins/credentials/CredentialsStoreAction/DomainWrapper/newCredentials.jelly
@@ -25,57 +25,59 @@
  -->
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:l="/lib/layout"
          xmlns:f="/lib/form" xmlns:c="/lib/credentials" xmlns:t="/lib/hudson">
-  <l:layout title="${%New Credentials}" norefresh="true" permission="${it.parent.CREATE}">
-    <st:include page="sidepanel.jelly"/>
-    <l:main-panel>
-      <j:set var="descriptor" value="${it.credentialDescriptor}"/>
-      <j:set var="instance" value="${null}"/>
-      <f:form action="createCredentials" method="POST" name="newCredentials">
-        <j:set var="descriptors" value="${it.store.credentialsDescriptors}" />
-        <j:choose>
-          <j:when test="${descriptors.size()==1}">
-            <f:entry title="${%Kind}" help="${descriptor.getHelpFile('credentials')}">
-              <i>${descriptors[0].displayName}</i>
-            </f:entry>
-            <f:rowSet name="credentials">
-              <j:set var="descriptor" value="${descriptors[0]}"/>
-              <j:set var="instance" value="${null}"/>
-              <f:class-entry descriptor="${descriptor}"/>
-              <st:include from="${descriptor}" page="${descriptor.configPage}"/>
-            </f:rowSet>
-          </j:when>
-          <j:otherwise>
-            <j:local>
-              <j:set var="it" value="${it.store}"/>
-              <!-- TODO revert to dropdownDescriptorSelector when baseline is 1.645+ which supports `capture` attribute -->
-              <f:dropdownList name="credentials" title="${%Kind}" help="${descriptor.getHelpFile('credentials')}">
-                <j:set var="current" value="${instance[credentials]}"/>
-                <j:set var="current" value="${current!=null ? current : null}"/>
-                <j:forEach var="descriptor" items="${descriptors}" varStatus="loop">
-                  <f:dropdownListBlock value="${loop.index}" title="${descriptor.displayName}"
-                                       selected="${current.descriptor==descriptor || (current==null and descriptor==attrs.default)}"
-                                       staplerClass="${descriptor.clazz.name}"
-                                       lazy="descriptor,it">
-                    <l:ajax>
-                      <j:set var="instance" value="${current.descriptor==descriptor ? current : null}"/>
-                      <st:include from="${descriptor}" page="${descriptor.configPage}"/>
-                    </l:ajax>
-                  </f:dropdownListBlock>
-                </j:forEach>
-              </f:dropdownList>
-              <!--f:dropdownDescriptorSelector field="credentials" title="${%Kind}" lazy="it"/-->
-            </j:local>
-          </j:otherwise>
-        </j:choose>
+  <j:if test="${it.credentialsModifiable}">
+    <l:layout title="${%New Credentials}" norefresh="true" permission="${it.parent.CREATE}">
+      <st:include page="sidepanel.jelly"/>
+      <l:main-panel>
+        <j:set var="descriptor" value="${it.credentialDescriptor}"/>
+        <j:set var="instance" value="${null}"/>
+        <f:form action="createCredentials" method="POST" name="newCredentials">
+          <j:set var="descriptors" value="${it.store.credentialsDescriptors}" />
+          <j:choose>
+            <j:when test="${descriptors.size()==1}">
+              <f:entry title="${%Kind}" help="${descriptor.getHelpFile('credentials')}">
+                <i>${descriptors[0].displayName}</i>
+              </f:entry>
+              <f:rowSet name="credentials">
+                <j:set var="descriptor" value="${descriptors[0]}"/>
+                <j:set var="instance" value="${null}"/>
+                <f:class-entry descriptor="${descriptor}"/>
+                <st:include from="${descriptor}" page="${descriptor.configPage}"/>
+              </f:rowSet>
+            </j:when>
+            <j:otherwise>
+              <j:local>
+                <j:set var="it" value="${it.store}"/>
+                <!-- TODO revert to dropdownDescriptorSelector when baseline is 1.645+ which supports `capture` attribute -->
+                <f:dropdownList name="credentials" title="${%Kind}" help="${descriptor.getHelpFile('credentials')}">
+                  <j:set var="current" value="${instance[credentials]}"/>
+                  <j:set var="current" value="${current!=null ? current : null}"/>
+                  <j:forEach var="descriptor" items="${descriptors}" varStatus="loop">
+                    <f:dropdownListBlock value="${loop.index}" title="${descriptor.displayName}"
+                                         selected="${current.descriptor==descriptor || (current==null and descriptor==attrs.default)}"
+                                         staplerClass="${descriptor.clazz.name}"
+                                         lazy="descriptor,it">
+                      <l:ajax>
+                        <j:set var="instance" value="${current.descriptor==descriptor ? current : null}"/>
+                        <st:include from="${descriptor}" page="${descriptor.configPage}"/>
+                      </l:ajax>
+                    </f:dropdownListBlock>
+                  </j:forEach>
+                </f:dropdownList>
+                <!--f:dropdownDescriptorSelector field="credentials" title="${%Kind}" lazy="it"/-->
+              </j:local>
+            </j:otherwise>
+          </j:choose>
 
-        <f:bottomButtonBar>
-          <f:submit value="${%OK}"/>
-        </f:bottomButtonBar>
-      </f:form>
-      <script>
-        // TODO remove this JENKINS-24662 workaround when baseline core has fix for root cause
-        window.setTimeout(function(){layoutUpdateCallback.call();}, 1000);
-      </script>
-    </l:main-panel>
-  </l:layout>
+          <f:bottomButtonBar>
+            <f:submit value="${%OK}"/>
+          </f:bottomButtonBar>
+        </f:form>
+        <script>
+          // TODO remove this JENKINS-24662 workaround when baseline core has fix for root cause
+          window.setTimeout(function(){layoutUpdateCallback.call();}, 1000);
+        </script>
+      </l:main-panel>
+    </l:layout>
+  </j:if>
 </j:jelly>
