+++ src/main/resources/com/cloudbees/plugins/credentials/CredentialsStoreAction/DomainWrapper/sidepanel.jelly
@@ -28,7 +28,9 @@
     <l:tasks>
     <j:set var="url" value="${h.getNearestAncestorUrl(request,it)}"/>
     <l:task icon="images/24x24/up.png" href="../.." title="${%Back to credential domains}" contextMenu="false"/>
+      <j:if test="${it.parent.credentialsModifiable}">
       <l:task icon="plugin/credentials/images/24x24/new-credential.png" href="newCredentials" title="${%Add Credentials}" permission="${it.parent.CREATE}" />
+      </j:if>
       <j:if test="${!it.global and it.parent.domainsModifiable}">
       <l:task icon="images/24x24/setting.png" href="${url}/configure" title="${%Configure}" permission="${it.parent.MANAGE_DOMAINS}"/>
       <l:task icon="images/24x24/edit-delete.png" href="${url}/delete" title="${%Delete domain}" permission="${it.parent.MANAGE_DOMAINS}"/>
